FastSpringBoneBufferBuilder を FastSpringBoneBuffer.FastSpringBoneBuffer に合体
